# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Agendamentos",
        "description": "Documentação da API de Agendamento de Consultas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Principal"
        }
    ],
    "paths": {
        "/appointments": {
            "get": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Listar consultas do usuário",
                "operationId": "c68dab577f29c04c4422e6cb86fa36ce",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filtrar por data (AAAA-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-05-15"
                        }
                    },
                    {
                        "name": "professional_id",
                        "in": "query",
                        "description": "Filtrar por ID do profissional",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Agendar uma consulta",
                "operationId": "0cfd7a3482605b745b634d591966d8bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "professional_id",
                                    "date",
                                    "time"
                                ],
                                "properties": {
                                    "professional_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-15"
                                    },
                                    "time": {
                                        "type": "string",
                                        "example": "10:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consulta agendada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Erros de validação ou conflito de horário"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{appointment}": {
            "get": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Obter detalhes da consulta",
                "operationId": "9b789941f10bf8dd3dea74aa7d6bd17c",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "description": "ID da Consulta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Consulta não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Cancelar uma consulta",
                "operationId": "d2e4aef8da50e5d82f441ed65ea4700d",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "description": "ID da Consulta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consulta cancelada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Consulta não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registrar um novo usuário",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@exemplo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso"
                    },
                    "422": {
                        "description": "Erros de validação"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Fazer login e obter o token",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@exemplo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    },
                    "422": {
                        "description": "Erros de validação"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Fazer logout do usuário",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Obter detalhes do usuário autenticado",
                "operationId": "99e9c2c91427f6bf63c2f0ef765a768f",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/professionals/{professional}/availability": {
            "get": {
                "tags": [
                    "Profissionais"
                ],
                "summary": "Obter horários disponíveis de um profissional",
                "operationId": "37a98d8f2ffdfe23b4f40b150ba5b2fc",
                "parameters": [
                    {
                        "name": "professional",
                        "in": "path",
                        "description": "ID do Profissional",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Data no formato AAAA-MM-DD",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-05-15"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "404": {
                        "description": "Profissional não encontrado"
                    }
                }
            }
        },
        "/professionals": {
            "get": {
                "tags": [
                    "Profissionais"
                ],
                "summary": "Obter lista de profissionais",
                "operationId": "57b34984b4212f4e8ca1c34c4ba4d230",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    }
                }
            },
            "post": {
                "tags": [
                    "Profissionais"
                ],
                "summary": "Cadastrar um novo profissional",
                "operationId": "2b6d4f766a9738c10b8fdbf20589f9fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "specialty"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dr. Carlos Silva"
                                    },
                                    "specialty": {
                                        "type": "string",
                                        "example": "Ortopedista"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profissional cadastrado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Erros de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/professionals/{professional}": {
            "get": {
                "tags": [
                    "Profissionais"
                ],
                "summary": "Obter detalhes de um profissional",
                "operationId": "ecb1bf8f236962bf01d82f4a551b1692",
                "parameters": [
                    {
                        "name": "professional",
                        "in": "path",
                        "description": "ID do Profissional",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "404": {
                        "description": "Profissional não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Consultas",
            "description": "Consultas"
        },
        {
            "name": "Autenticação",
            "description": "Autenticação"
        },
        {
            "name": "Profissionais",
            "description": "Profissionais"
        }
    ]
}